backend/requirements.txt
Django==5.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
python-dotenv==1.0.1
mysqlclient==2.2.1
django-cors-headers==4.3.1
python-dateutil==2.8.2
pytz==2024.1
google-genai==0.3.0
moviepy==1.0.3
SpeechRecognition==3.10.4
pydub==0.25.1
opencv-python==4.9.0.80
celery==5.3.6
redis==5.0.3
django-celery-beat==2.6.0
django-celery-results==2.6.0
assemblyai==0.48.4
openai-whisper
ffmpeg-python==0.2.0
